struct: rename files profile -> analyze
